requirements.txt
Django==3.2.4
django-multiupload==0.10.0
